Complete #13: Computing dice expressions into single numbers within strings
Adds ReplaceWrappedExpressions<T> to Dice, along with documentation on how it can be used.
Here's a summary here anyway:
It requires a string to be passed, returns that string with all expressions replaced with one single object of type T.
It can take custom wrappers for expressions (both ending and beginning),
and escaping beginning can be done with an optional character, or not at all. (potentially faster, at the cost of being able to have that character in your returned string.

Adds SetupToEval() to DiceTests because I needed a unified way to use a working evaluator.